Added a really basic help window
git-svn-id: https://vimpc.svn.sourceforge.net/svnroot/vimpc/trunk@67 fc90b29e-3040-4360-8be2-b397c2834534

+++ src/scrollwindow.cpp
@@ -115,7 +115,7 @@ bool ScrollWindow::AutoScroll() const
 }
 
 
-uint16_t ScrollWindow::FirstLine() const
+uint32_t ScrollWindow::FirstLine() const
 {
    uint16_t result = 0;
